Support running in individual receive mode
Modify to add a build flag SEND_BACK which controls whether the
two board configuration is used for this example or whether a
single board in receive mode, which can be a host computer, is
used as the transmitter instead.  When SEND_BACK is not defined
the code related to sending back the transmitted bytes are not
sent back and instead the code continues to recieve, counts bytes
received, and toggles LED2 each time a buffer is received.

With this implementation a single send of a file containing the
string defined in aTxBuffer completes successfully, and the led
toggles each time it's sent.

When a larger file is sent, the led toggles continuously and
no error LED is displayed