Add interrupt driven GPIO flow control for RTS
Since hardware flow control is not asserting the RTS soon enough,
add my own hack which sets/clears flow control based on fifo threshold.

With this change I can simulate processing delays of 10-15ms during
buffer transfers, and I don't drop any bytes, even when the duration
of the wait is longer than the approx 10ms it takes to complete the
transfer of the UART buffer.